Added two sptpol newdat files that are estimates of bandpowers for 2012 111 sq degrees with two different calibration errors.
Also made changes to fisher_utils.py to allow for multiple Dl realizations at once.

Removed a squared factor in the banderror sum term which I believe was incorrect.  I want to add bandpower errors in an "observation" sense,
not a quadrature sum sense.  Each of the errors is a "measurement" of that bandpower's error, so they should average down.